% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_treatment_effect}
\alias{estimate_treatment_effect}
\title{Title}
\usage{
estimate_treatment_effect(outcome, treatment)
}
\arguments{
\item{outcome}{numeric value for outcome measure at final time point}

\item{treatment}{categorical variable for treatment arm}
}
\value{
A list with:
\describe{
  \item{treated_mean}{Mean outcome for treated group.}
  \item{control_mean}{Mean outcome for control group.}
  \item{ate}{Average Treatment Effect (treated_mean - control_mean).}
}
}
\description{
Title
}
\examples{
set.seed(123)
y <- rnorm(100, mean = 5) + rbinom(100, 1, 0.5) * 2
treat <- rbinom(100, 1, 0.5)
estimate_treatment_effect(y, treat)
}
