% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-info.R
\name{print.bcmpapers}
\alias{print.bcmpapers}
\title{print.bcmpapers}
\usage{
\method{print}{bcmpapers}(x)
}
\arguments{
\item{x}{an object of class bcmpapers}
}
\description{
It prints the list of papers offered by BCM college per semester.
}
