% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-info.R
\name{print.bcmprogram}
\alias{print.bcmprogram}
\title{print.bcmprogram}
\usage{
\method{print}{bcmprogram}(x)
}
\arguments{
\item{x}{an object of class bcmprogram}
}
\description{
It prints the list of programs offered by the college.
}
