% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-info.R
\name{print.bcmresult}
\alias{print.bcmresult}
\title{print.bcmresult}
\usage{
\method{print}{bcmresult}(x)
}
\arguments{
\item{x}{an object of class bcmresult}
}
\description{
It print the result for each year every year.
}
